Works in Mac, does not work in Linux/Jutge
- ADD immersive functions, it is better like this because this way, if I
were to call these functions in another place, I would not have to
remember which previous initializations I would need for it.
- MODIFY crear_cuadro_resultado_final to a proper recursive function.
- Note: Everything works if executed in Mac, but in Linux (and Jutge),
cin >> string works different. In Mac, it reads everything until an
endline, but in Linux, it reads everything until it finds a space.
To fix this, change leer_cuadro_resultado_matches().